(ValueParser) Allow for multiline values in CSV
Some values in my CSV files contained contained linefeed or carriage return characters (within a field, not just at the end of lines). When trying to work with these fields with the ValueReplacer plugin, they did not get replaced. I fixed this with the changes on this commit by allowing multiline wildcard (.) matches in 3 preg_ regex functions (by adding the 's' after the final pattern delimiter).