BUG FIX Image label override
Originally, when the same image is added to gallery more than once, if
image label is not specified,  the label will be updated with null, as
result it resets image labels.

This has 2 downsides:
1. It resets any manually saved labels through Magento backend.
2. This is not good with the wildcard feature. Because magmi handles
images in this order (I believe): base image, gallery, small image and
at last thumbnail, the label of base image will be reset to null when
gallery contains wildcard path.

This fix basically checks if image label is not specified, do not update
the label to null.