NEW FEATURE Auto assign images to configurable
If this feature is turned on, it will assign the base image of the first
associated simple to be the base image, thumbnail and small image of the
configurable, AND add all gallery images of associated simples to the
configurable.

'back' image label support is added. If both features are turned on  ,
When auto adding images to the gallery of configurable, it will clear
any existing 'back' label and set the base image of the second simple to
'back'. The labels of simple products will not be affected.

Options are added to switch both features on and off.